Correction to radar location
Corrects an error in the hdw.dat.sto file discovered by Jim Wild which located the radar 1 deg west of its actual location.  Unclear where the error in the file originated from and how long it had been in place, but it is possible it was a typo in the original file.